+++ src/assets/images/index.ts
@@ -4,6 +4,7 @@ import dashboard2 from "../images/dashboard2.png";
 import dashboard3 from "../images/dashboard3.png";
 import dashboard4 from "../images/dashboard4..png";
 import verified from "../images/verified.png";
+import NepalFlag from "../images/flag-nepal.png";
 export const images = {
   logo,
   dashboard1,
@@ -11,4 +12,5 @@ export const images = {
   dashboard3,
   dashboard4,
   verified,
+  NepalFlag,
 };